pickrst_neon: add missing cast / quiet warning w/cl.exe
fixes:
av1\encoder\arm\neon\pickrst_neon.c(227): warning C4244: 'return':
conversion from 'uint64_t' to 'uint8_t', possible loss of data

Bug: b/277255390

+++ av1/encoder/arm/neon/pickrst_neon.c
@@ -224,7 +224,8 @@ static INLINE uint8_t find_average_neon(const uint8_t *src, int src_stride,
       h_limit += h_overflow;
       h_limit = height > h_overflow ? h_overflow : height;
     } while (h < height);
-    return (horizontal_long_add_u32x2(avg_u32) + sum) / (width * height);
+    return (uint8_t)((horizontal_long_add_u32x2(avg_u32) + sum) /
+                     (width * height));
   }
   int i = height;
   do {
